requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
colorama==0.4.4
coverage==6.4.1
decorator==5.1.1
Django==3.1.3
django-autoslug==1.9.8
django-debug-toolbar==3.1.1
django-enum-choices==2.1.4
django-model-utils==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.1.0
Faker==13.13.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipython==7.19.0
jedi==0.18.1
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.29
psycopg2==2.8.6
py==1.11.0
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.2.2.post1
typing-extensions==4.2.0
wcwidth==0.2.5
zipp==3.8.0
